+++ components/view_in_browser.go
@@ -2,10 +2,10 @@ package components
 
 type ViewInBrowser struct {
 	Id   ComponentId `json:"id"`
-	Text string      `json:"text"`
+	Text *string     `json:"text"`
 }
 
-func NewViewInBrowser(text string) *ViewInBrowser {
+func NewViewInBrowser(text *string) *ViewInBrowser {
 	return &ViewInBrowser{
 		Id:   ComponentIdViewInBrowser,
 		Text: text,